parsing JSON int and double values as int and double. store_def.json is changed to support this.

+++ SoomlaAndroidStore/src/com/soomla/store/StoreInfo.java
@@ -90,7 +90,7 @@ public class StoreInfo {
                         jsonNode.path("image").textValue(),
                         jsonNode.path("itemId").textValue(),
                         gItem,
-                        jsonNode.path("price").intValue(),
+                        jsonNode.path("price").doubleValue(),
                         jsonNode.path("amount").intValue());
 
                 mPacksOptions.put(jsonNode.path("productId").textValue(), pack);
